refactor: serve command
