+++ packages/serve/src/index.ts
@@ -1,24 +1,29 @@
 import startDevServer from './startDevServer';
 
 class ServeCommand {
-    async apply(cli): Promise<void> {
+    apply(cli): void {
         const { program, logger } = cli;
+
+        try {
+            // eslint-disable-next-line node/no-extraneous-require
+            require('webpack-dev-server');
+        } catch (err) {
+            logger.error(`You need to install 'webpack-dev-server' for running 'webpack serve'.\n${err}`);
+            process.exit(2);
+        }
+
+        // TODO
+        // eslint-disable-next-line node/no-extraneous-require
+        const devServerFlags = require('webpack-dev-server/bin/cli-flags').devServer;
 
         program
             .command('serve')
             .alias('s')
             .description('Run the webpack Dev Server')
             .usage('serve [options]')
+            .option('--port <value>')
             .action(async (program) => {
                 const options = program.opts();
-
-                try {
-                    // eslint-disable-next-line node/no-extraneous-require
-                    require('webpack-dev-server');
-                } catch (err) {
-                    logger.error(`You need to install 'webpack-dev-server' for running 'webpack serve'.\n${err}`);
-                    process.exit(2);
-                }
 
                 // // Add WEBPACK_SERVE environment variable
                 // if (webpackArgs.env) {
@@ -34,7 +39,7 @@ class ServeCommand {
 
                 const compiler = await cli.getCompiler({});
 
-                await startDevServer(compiler, {});
+                await startDevServer(compiler, options);
             });
     }
 }
